+++ src/utils.js
@@ -1,0 +1,4 @@
+import { parseFunctionFromExpression } from 'enketo-core/src/js/utils';
+import addXPathExtensionsOc from 'enketo-xpath-extensions-oc';
+
+module.exports = { parseFunctionFromExpression, addXPathExtensionsOc };